chore: rename Client to ApiClient
